CI green is possible again: two lints, and the Windows link that never worked
The clippy half was mine: a useless u32 conversion in the presenter and an
unsafe block the macOS adapter no longer needs.

The Windows half was chronic, not a regression -- no CI run has been green
in the last forty. `cargo test --workspace --lib` links krate_runtime's
debug test binary, and whisper's cmake objects only agree with the
workspace's mandatory static CRT (K-037: without it krate.exe silently
fails to start on clean Windows 11) in the release profile; the debug link
dies on unresolved __imp_ stdio symbols, and always has, while every
release Windows build links the identical code without complaint. The
Windows job now runs the same tests under the profile that actually links.

+++ crates/presenter-gpu/src/lib.rs
@@ -123,7 +123,7 @@ fn draw_layout(scene: &mut vello::Scene, layout: &Layout<()>, color: u32, x: f32
                     let gy = y + run_y - g.y;
                     run_x += g.advance;
                     vello::Glyph {
-                        id: u32::from(g.id),
+                        id: g.id,
                         x: gx,
                         y: gy,
                     }
